// src/types.ts
export type Language = 'en' | 'ar' | 'de' | 'es';
export type Direction = 'ltr' | 'rtl';
export type Platform = 'GetYourGuide' | 'Viator';

export interface NavLink {
    type: 'link';
    href: string;
    text: string;
}

export interface NavDropdown {
    type: 'dropdown';
    title: string;
    items: {
        href: string;
        text: string;
        description: string;
    }[];
}

export interface HeaderContent {
    navItems: (NavLink | NavDropdown)[];
    cta: string;
}